endpoints terminados Order y OrderItem

+++ src/main/java/com/uade/tpo/marketplace/entity/Order.java
@@ -21,7 +21,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 public class Order {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id_order;
+    private Long id_order;
     
     @Column
     private String payment_method;
